Refactor property setting
Previously, various utility overloads of `Object::Set()` were provided.
This was problematic, because accidental unexpected conversions
could happen (e.g. `const char*` to `bool`) and overloads were
ambiguous (e.g. `int` could be cast to `double` and `bool` and therefore
could not be passed directly).

To address this, pass a template parameter instead, and introduce
a helper `Value::From` method that could create a value based on the
type of object it receives, and that uses clear precedence rules for
determining the kind of JS value to create.

Fixes: https://github.com/nodejs/node-addon-api/issues/86